+++ src/controllers/users.js
@@ -33,4 +33,28 @@ module.exports = {
       next({ status: 400, send: { msg: 'Usuario no creado', err: error } });
     }
   },
+  put: async (req, res, next) => {
+    try {
+      const { id } = req.params;
+      let user = await Users.findByIdAndUpdate(id, req.body);
+      next({
+        status: 201,
+        send: { msg: 'Usuario actualizado', data: req.body },
+      });
+    } catch (error) {
+      next({
+        status: 400,
+        send: { msg: 'Usuario no actualizado', data: error },
+      });
+    }
+  },
+  delete: async (req, res, next) => {
+    try {
+      const { id } = req.params;
+      let user = await Users.findByIdAndDelete(id);
+      next({ status: 201, send: { msg: 'Usuario eliminado' } });
+    } catch (error) {
+      next({ status: 400, send: { msg: 'Usuario no eliminado', data: error } });
+    }
+  },
 };
