+++ src/routes/users.js
@@ -1,12 +1,14 @@
-// const express = require('express');
-// const { route } = require('.');
-// const router = express.Router();
-// const userController = require('../controllers/users');
+const express = require('express');
+const router = express.Router();
+const users = './src/db/user.json';
+const fs = require('fs');
+const userController = require('../controllers/users');
 
-// router.get('/', userController.getAll);
-// router.get('/:id', userController.getById);
-// router.post('/', userController.post);
-// router.put('/:id', userController.put);
-// router.delete('/:id', userController.delete);
+router.get('/', userController.getAll);
 
-// module.exports = router;
+router.get('/:id', userController.getById);
+
+router.post('/', userController.post);
+router.put('/:id', userController.put);
+router.delete('/:id', userController.delete);
+module.exports = router;
